사다리게임 결과맵 문자열 반환의 책임을 LadderGameService에서 지도록 함.
- Ladder getLadder -> String getResultMap() : 사다리를 반환하지 말고, 결과 문자열을 반환하도록 함.
- OutputView에서는 결과 문자열을 출력하는 역할만 줌.
