+++ src/main/java/domain/game/LadderGameService.java
@@ -1,9 +1,7 @@
 package domain.game;
 
-import domain.ladder.Ladder;
-
 public interface LadderGameService {
 
     void initLadderGame(int numberOfUsers, int height);
-    Ladder getLadder();
+    String getResultMap();
 }
