+++ src/main/java/view/output/OutputView.java
@@ -1,9 +1,7 @@
 package view.output;
 
-import domain.ladder.Ladder;
-
 public interface OutputView {
 
-    void printLadder(Ladder ladder);
+    void printResultMap(String resultMap);
 
 }
